backend/pyproject.toml
[tool.poetry]
name = "kursovaya"
version = "0.1.0"
description = ""
authors = ["Arslan Semenov IB-234B"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
